requirements.txt
pytest==9.0.2
httpx==0.28.1
jsonschema==4.25.1
Faker==38.2.0
pydantic==2.12.5
pydantic-settings==2.12.0
allure-pytest==2.15.2
email-validator==2.3.0
uuid==1.30
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
allure-python-commons==2.15.2
swagger-coverage-tool==0.31.0